JENKINS-45555 : Disable spy through environment when running maven-invoker IT tests
By default, IT tests ran by the maven-invoker-plugin will see the
JAVA_TOOL_OPTIONS environment variable and thus enable the event-spy.
All IT tests runs will be recorded and their artefacts will be archived
and fingerprinted as they were from the main project. But one do not
care about that artefacts, they came from test projets.
To avoid this, if we detect the launch of the maven-invoker-plugin:run
goal, we append to its configuration the environment variable that will
disable the event-spy during IT tests.